Create Jwt Auth

+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\AuthController;
 use App\Http\Controllers\CategoryController;
 use Illuminate\Support\Facades\Route;
 
@@ -8,10 +9,16 @@ Route::get('/', function () {
 });
 
 Route::controller(CategoryController::class)->group(function () {
-    Route::get(config("endpoint.CATEGORY"), "getCategoryWithPagination");
-    Route::get(config("endpoint.CATEGORY") . "/all", "getCategoryWithoutPagination");
-    Route::get(config("endpoint.CATEGORY") . "/{categoryId}", "getCategoryById");
-    Route::post(config("endpoint.CATEGORY"), "saveCategory");
-    Route::put(config("endpoint.CATEGORY") . "/{categoryId}", "updateCategory");
-    Route::delete(config("endpoint.CATEGORY") . "/{categoryId}", "deleteCategory");
+    Route::get(config("endpoint.CATEGORY"), "getCategoryWithPagination")->middleware('auth');
+    Route::get(config("endpoint.CATEGORY") . "/all", "getCategoryWithoutPagination")->middleware('auth');
+    Route::get(config("endpoint.CATEGORY") . "/{categoryId}", "getCategoryById")->middleware('auth');
+    Route::post(config("endpoint.CATEGORY"), "saveCategory")->middleware('auth');
+    Route::put(config("endpoint.CATEGORY") . "/{categoryId}", "updateCategory")->middleware('auth');
+    Route::delete(config("endpoint.CATEGORY") . "/{categoryId}", "deleteCategory")->middleware('auth');
+});
+
+Route::controller(AuthController::class)->group(function () {
+    Route::post(config("endpoint.AUTH_LOGIN"), "login");
+    Route::post(config("endpoint.AUTH_LOGOUT"), "logout");
+    Route::get(config("endpoint.AUTH_USER"), "getUserData");
 });
